+++ src/main/java/io/github/duckysmacky/utilimate/collections/MapUtils.java
@@ -4,6 +4,7 @@ import io.github.duckysmacky.utilimate.enums.Order;
 
 import java.util.*;
 import java.util.Map.Entry;
+import java.util.stream.Stream;
 
 /** <p> Utility class providing static methods for common operations on maps.
  * Includes methods for filtering, sorting by key or value, merging maps, and more.
@@ -60,5 +61,20 @@ public final class MapUtils {
         Map<V, K> invertedMap = new HashMap<>();
         for (Entry<K, V> entry : entryList) invertedMap.put(entry.getValue(), entry.getKey());
         return invertedMap;
+    }
+
+    /** Merges multiple maps together by adding all the entries into a single one (assuming all the keys are unique,
+     * else overrides old key)
+     * @param maps maps to merge
+     * @param <K> key type
+     * @param <V> value type
+     * @return {@code Map} merged map
+     * @since 0.2.0
+     */
+    @SafeVarargs
+    public static <K, V> Map<K, V> merge(Map<K, V>... maps) {
+        Map<K, V> mergedMap = new HashMap<>();
+        for (Map<K, V> map : maps) mergedMap.putAll(map);
+        return mergedMap;
     }
 }